Make calculation of delta_temp type safe by cast

+++ lib/Instructions.c
@@ -731,11 +731,11 @@ void read_initial_settings(struct All_variables *E)
       input_float("start_temp",&(E->composition.start_temp),"0,0,nomax",m);
       input_float("end_temp",&(E->composition.end_temp),"1,0,nomax",m);
       input_int("ntdeps",&(E->composition.ntdeps),"1000,0,nomax",m);
-      E->composition.delta_temp = (E->composition.end_temp - E->composition.start_temp) / E->composition.ntdeps;
+      E->composition.delta_temp = (E->composition.end_temp - E->composition.start_temp) / (float) E->composition.ntdeps;
   }
   else {
        E->composition.ntdeps = 1000;
-       E->composition.delta_temp = E->data.ref_temperature / E->composition.ntdeps;
+       E->composition.delta_temp = E->data.ref_temperature / (float) E->composition.ntdeps;
   }
   if (E->composition.end_temp < 1e-7) E->composition.end_temp = E->data.ref_temperature;
