+++ module/setProperties.c
@@ -672,10 +672,10 @@ PyObject * pyCitcom_Tracer_set_properties(PyObject *self, PyObject *args)
          getFloatProperty(properties, "end_temp", E->composition.end_temp, fp);
          getIntProperty(properties, "ntdeps", E->composition.ntdeps, fp);
          E->composition.delta_temp = (E->composition.end_temp - E->composition.start_temp)
-                                        / E->composition.ntdeps;
+                                        / (float) E->composition.ntdeps;
     } else {
          E->composition.ntdeps = 1000;
-         E->composition.delta_temp = E->data.ref_temperature / E->composition.ntdeps;
+         E->composition.delta_temp = E->data.ref_temperature / (float) E->composition.ntdeps;
     }
     if (E->composition.end_temp < 1e-7) E->composition.end_temp = E->data.ref_temperature;
